+++ tests/test_variables_files/complex_types.tf
@@ -1,0 +1,15 @@
+variable "aws_credentials" {
+  type = map({
+    nonprod = map(string)
+    prod    = map(string)
+  })
+  default = {
+    nonprod = {
+      profile = "pretf-nonprod"
+    }
+
+    prod = {
+      profile = "pretf-prod"
+    }
+  }
+}